Extract Msf::DBManager::Import::Outpost24
MSP-11124
